changed overuse of size_t to unsigned int globally

+++ src/entities.h
@@ -9,18 +9,18 @@ struct Noble;
 struct City;
 
 typedef struct User {
-	size_t level;
-	size_t prowess;
-	size_t army;
-	size_t armycap;
-	size_t charisma;
-	int denars;
-	int renown;
-	int honor;
-	int kills;
-	int intown;
-	int x;
-	int y;
+	unsigned int level;
+	unsigned int prowess;
+	unsigned int army;
+	unsigned int armycap;
+	unsigned int charisma;
+	unsigned int denars;
+	unsigned int renown;
+	unsigned int honor;
+	unsigned int kills;
+	unsigned int intown;
+	unsigned int x;
+	unsigned int y;
 	char season[10];
 	char name[13]; 
 	char title[20];
@@ -28,43 +28,43 @@ typedef struct User {
 } User;
 
 typedef struct Buddy {
-	size_t level;
-	size_t prowess;
-	int inparty;
-	int x;
-	int y;
+	unsigned int level;
+	unsigned int prowess;
+	unsigned int inparty;
+	unsigned int x;
+	unsigned int y;
 	char name[13];
 	char location[20];
 } Buddy;
 
 typedef struct {
-	size_t size;
+	unsigned int size;
 	Buddy buddies[6];
 } BuddyList;
 
 typedef struct{
-	size_t maa;
-	size_t maaupkeep;
-	size_t pspear;
-	size_t pspearupkeep;
-	size_t pbow;
-	size_t pbowupkeep;
-	size_t total;
-	size_t totalupkeep;
+	unsigned int maa;
+	unsigned int maaupkeep;
+	unsigned int pspear;
+	unsigned int pspearupkeep;
+	unsigned int pbow;
+	unsigned int pbowupkeep;
+	unsigned int total;
+	unsigned int totalupkeep;
 	char maaname[20];
 	char pspearname[20];
 	char pbowname[20];
 } PartyList;
 
 typedef struct City {
-	size_t pop;
-	size_t danger;
-	size_t wealth;
-	int iscity;
-	int x;
-	int y;
-	int hasdel;
-	int hassla;
+	unsigned int pop;
+	unsigned int danger;
+	unsigned int wealth;
+	unsigned int iscity;
+	unsigned int x;
+	unsigned int y;
+	unsigned int hasdel;
+	unsigned int hassla;
 	char name[10];
 	char location[30];
 	char gm[10];
@@ -73,26 +73,26 @@ typedef struct City {
 } City;
 
 typedef struct {
-	size_t size;
+	unsigned int size;
 	City* cities[12];
 } CityList;
 
 typedef struct Noble{
-	size_t totalfiefs;
-	int relations;
+	unsigned int totalfiefs;
+	unsigned int relations;
 	City* fiefs[3];
 	char name[20];
 } Noble;
 
 typedef struct {
-	size_t size;
+	unsigned int size;
 	Noble* nobles[7];
 } NobleList;
 
 typedef struct {
-	int questid;
-	size_t quantity;
-	size_t quality;
+	unsigned int questid;
+	unsigned int quantity;
+	unsigned int quality;
 	Noble* recipient1;
 	City* recipient2;
 	char info[50];
@@ -100,29 +100,29 @@ typedef struct {
 } Item;
 
 typedef struct {
-	size_t size;
+	unsigned int size;
 	Item items[10];	
 } Inventory;
 
 typedef struct {
-	int relation_buff;
-	int renown_gain;
-	int questid;
+	unsigned int relation_buff;
+	unsigned int renown_gain;
+	unsigned int questid;
 	Noble* giver;
 	Noble* target;
 } Quest1;
 
 typedef struct {
-	int to_kill;
-	int relation_buff;
-	int renown_gain;
-	int questid;
+	unsigned int to_kill;
+	unsigned int relation_buff;
+	unsigned int renown_gain;
+	unsigned int questid;
 	City* giver;
 } Quest2;
 
 typedef struct {
-	size_t totaldel;
-	size_t totalsla;
+	unsigned int totaldel;
+	unsigned int totalsla;
 	Quest1 deliveries[5];
 	Quest2 slayings[5];
 } QuestList;
